backend/requirements.txt
flask
flask-sqlalchemy
flask-cors
flask-socketio
flask-jwt-extended
pymysql
python-dotenv
requests
bcrypt
pydantic
redis
